requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==3.1.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
email-validator==2.0.0.post2
Pillow==10.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
requests==2.31.0
